scripts/transcribe/pyproject.toml
[project]
name = "transcribe"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flash-attn>=2.8.3",
    "python-telegram-bot>=22.6",
    "qwen-asr",
    "soundfile>=0.13.1",
    "torch>=2.9.1",
]

[tool.uv]
dev-dependencies = [
    "pyright>=1.1.399",
]

[tool.uv.workspace]
members = [
    "Qwen3-ASR",
]

[tool.uv.sources]
qwen-asr = { workspace = true, editable = true }
